backend/requirements.txt
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9        # PostgreSQL driver
alembic==1.13.1               # Migraciones
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4        # Password hashing
python-multipart==0.0.6       # File uploads
###pillow###==10.3.0                # Image processing
python-dotenv==1.0.0

